add: splitting code  into multilayers - deleteTodo
